// src/types/params.ts
export interface QueryParams {
  lastName?: string;
  firstName?: string;
  informalName?: string;
  registrationNumber?: string;
  registrationClass?: string;
  registrationStatus?: string;
  contactLensMentor?: string;
  practiceName?: string;
  cityOrTown?: string;
  postalCode?: string;
  languageOfService?: string;
  areaOfService?: string;
}